% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_door.R
\name{change_door}
\alias{change_door}
\title{Change Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical indicating if the player wants to stay with the first pick}

\item{opened.door}{An integer representing the number of the opened door}

\item{a.pick}{An integer representing the number of the selected door}
}
\value{
An integer representing the number of the final pick
}
\description{
This function takes the 'stay' (logical), 'opened.door' and 'a.pick' (selected door) as inputs.
If the stay is TRUE, it returns the selected door.
If the stay is FALSE, it changes the selected door to the other unopened door.
}
\examples{
change_door(FALSE, 2, 1)

}
