% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_winner.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer representing the final selected door}

\item{game}{A vector of "goat" and "car" representing the game}
}
\value{
A string indicating whether the player wins or loses
}
\description{
This function takes the 'final.pick' and 'game' as inputs
It checks if the element at the 'final.pick' index of the 'game' vector is "car" or "goat"
If it's "car" the function returns "WIN" else it returns "LOSE"
}
\examples{
determine_winner(2, c("goat", "car", "goat"))

}
