% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{mcstate_domain_expand}
\alias{mcstate_domain_expand}
\title{Expand (and check) domain against a packer}
\usage{
mcstate_domain_expand(domain, packer)
}
\arguments{
\item{domain}{A two-column matrix as defined in \link{mcstate_model},
with row names corresponding to either logical names (e.g., \code{b})
or specific names \code{b[1]} that are present in your packer.
\code{NULL} is allowed where all parameters are defined over the
entire real line.}

\item{packer}{A \link{mcstate_packer} object}
}
\value{
A two dimensional matrix representing your domain, or
\code{NULL} if \code{domain} was given as \code{NULL}.
}
\description{
Check and expand a domain, where it is used alongside a
\link{mcstate_packer} object.  This can be used to expand domains for
logical parameters (e.g. a vector \code{b}) into its specific names
(e.g., \code{b[1]}, \code{b[2]}, etc) without having to rely on the
internals about how these names are constructed.
}
\examples{

packer <- mcstate_packer(c("a", "b"), list(x = 3, y = c(2, 2)))
mcstate_domain_expand(NULL, packer)
mcstate_domain_expand(rbind(x = c(0, 1)), packer)
mcstate_domain_expand(rbind(x = c(0, 1), "x[2]" = c(0, Inf)), packer)
mcstate_domain_expand(rbind(x = c(0, 1), "y" = c(0, Inf)), packer)
}
