% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{mcstate_dsl}
\alias{mcstate_dsl}
\title{Domain Specific Language for mcstate}
\usage{
mcstate_dsl(x, type = NULL)
}
\arguments{
\item{x}{The model as an expression.  This may be given as an
expression, as a string, or as a path to a filename.  Typically,
we'll do a reasonable job of working out what you've provided
but use the \code{type} argument to disambiguate or force a
particular interpretation.  The argument uses rlang's quosures
to allow you to work with expressions directly; see examples for
details.}

\item{type}{Force interpretation of the type of}
}
\value{
A \link{mcstate_model} object derived from the expressions you
provide.
}
\description{
Create a model using the mcstate2 DSL; this function will likely
change name in future, as will its interface.
}
\examples{

# Expressions that correspond to models can be passed in with no
# quoting
mcstate_dsl(a ~ Normal(0, 1))
mcstate_dsl({
  a ~ Normal(0, 1)
  b ~ Exponential(1)
})

# You can also pass strings
mcstate_dsl("a ~ Normal(0, 1)")
}
