% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl-distributions.R
\name{mcstate_dsl_distributions}
\alias{mcstate_dsl_distributions}
\title{Information about supported distributions}
\usage{
mcstate_dsl_distributions()
}
\value{
A data.frame with columns
\itemize{
\item \code{name} the name of the distribution; each name begins with a
capital letter, and there are duplicate names where different
parametrisations are supported.
\item \code{args} the arguments of all parameters, \emph{except} the random
variable itself which is given as the first argument to density
functions.
}

We may expand the output here in future to include information on
if distributions have support in C++, but we might end up
supporting everything this way soon.
}
\description{
Report information about supported distributions in the dsl.  This
is primarily intended for use in packages which use
\link{mcstate_dsl_parse_distribution}, as this function reports
information about which distributions and arguments would succeed
there.
}
