% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl-parse-error.R
\name{mcstate_dsl_error_explain}
\alias{mcstate_dsl_error_explain}
\title{Explain mcstate2 error}
\usage{
mcstate_dsl_error_explain(code)
}
\arguments{
\item{code}{The error code, as a string, in the form \code{Exxx} (a
capital "E" followed by three numbers)}
}
\value{
Nothing, this is called for its side effect only
}
\description{
Explain error codes produced by mcstate2.  This is a work in progress,
and we would like feedback on what is useful as we improve it.
The idea is that if you see an error you can link through to get
more information on what it means and how to resolve it.  The
current implementation of this will send you to the rendered
vignettes, but in the future we will arrange for offline rendering
too.
}
