% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mcstate_model_density}
\alias{mcstate_model_density}
\title{Compute log density}
\usage{
mcstate_model_density(model, parameters)
}
\arguments{
\item{model}{An \link{mcstate_model} object}

\item{parameters}{A vector or matrix of parameters}
}
\value{
A log-density value, or vector of log-density values
}
\description{
Compute log density for a model.  This is a wrapper around the
\verb{$density} property within an \link{mcstate_model} object.
}
\seealso{
\link{mcstate_model_gradient} for computing gradients and
\link{mcstate_model_direct_sample} for sampling from a model.
}
