% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{mcstate_runner_serial}
\alias{mcstate_runner_serial}
\title{Run MCMC chain in series}
\usage{
mcstate_runner_serial()
}
\value{
A runner of class \code{mcstate_runner} that can be passed to
\code{\link[=mcstate_sample]{mcstate_sample()}}
}
\description{
Run MCMC chains in series (one after another).  This is the
simplest chain runner, and the default used by \code{\link[=mcstate_sample]{mcstate_sample()}}.
It has nothing that can be configured (yet).
}
