% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{mcstate_sample_continue}
\alias{mcstate_sample_continue}
\title{Continue sampling}
\usage{
mcstate_sample_continue(samples, n_steps, restartable = FALSE)
}
\arguments{
\item{samples}{A \code{mcstate_samples} object created by
\code{\link[=mcstate_sample]{mcstate_sample()}}}

\item{n_steps}{The number of new steps to run}

\item{restartable}{Logical, indicating if the chains should be
restartable.  This will add additional data to the chains
object.}
}
\value{
A list of parameters and densities
}
\description{
Continue (restart) chains started by \link{mcstate_sample}.  Requires
that the original chains were run with \code{restartable = TRUE}.
Running chains this way will result in the final state being
exactly the same as running for the total (original + continued)
number of steps in a single push.
}
