% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler-hmc.R
\name{mcstate_sampler_hmc}
\alias{mcstate_sampler_hmc}
\title{Create HMC}
\usage{
mcstate_sampler_hmc(
  epsilon = 0.015,
  n_integration_steps = 10,
  vcv = NULL,
  debug = FALSE
)
}
\arguments{
\item{epsilon}{The step size of the HMC steps}

\item{n_integration_steps}{The number of HMC steps per step}

\item{vcv}{A variance-covariance matrix for the momentum vector.
The default uses an identity matrix.}

\item{debug}{Logical, indicating if we should save all
intermediate points and their gradients.  This will add a vector
"history" to the details after the integration.  This \emph{will}
slow things down though as we accumulate the history
inefficiently.}
}
\value{
A \code{mcstate_sampler} object, which can be used with
\link{mcstate_sample}
}
\description{
Create a Hamiltonian Monte Carlo sampler, implemented using the
leapfrog algorithm.
}
