% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler-mh.R
\name{mcstate_sampler_metropolis_hastings}
\alias{mcstate_sampler_metropolis_hastings}
\title{Metropolis-Hastings Sampler}
\usage{
mcstate_sampler_metropolis_hastings(proposal = NULL, vcv = NULL)
}
\arguments{
\item{proposal}{A proposal function; must take a vector of
parameters and produce a new vector of proposed parameters.}

\item{vcv}{A variance covariance matrix to generate a \code{proposal}
function from.  If you want multivariate Gaussian proposal, this
is likely simpler than supplying your own \code{proposal}, and
generally more efficient too.}
}
\value{
A \code{mcstate_sampler} object, which can be used with
\link{mcstate_sample}
}
\description{
Create a simple Metropolis-Hastings sampler (really just a plain
Metropolis sampler since we assume symmetric proposals and don't
yet support the correction for asymmetric proposal).
}
