% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl-differentiate-expr.R
\name{monty_differentiation}
\alias{monty_differentiation}
\title{Differentiate expressions}
\usage{
monty_differentiation()
}
\value{
A list of related objects:
\itemize{
\item \code{differentiate}: A function that can differentiate an expression
with respect to a variable (as a string).
\item \code{maths}: Some mathematical utilities for constructing
expressions.  This will be documented later, but the most useful
bits on here are the function elements \code{times}, \code{plus} and
\code{plus_fold}.
}

We will expand this soon to advertise what functions we are able
to differentiate to allow programs to fail fast.
}
\description{
Differentiate expressions in the monty DSL.  This function is
exported for advanced use, and really so that we can use it from
odin.  But it has the potential to be generally useful, so while
we'll tweak the interface quite a lot over the next while it is
fine to use if you can handle some disruption.
}
\details{
R already has support for differentiating expressions using \link{D},
which is useful for creating derivatives of simple functions to
pass into non-linear optimisation.  We need something a bit more
flexible for differentiating models in the monty DSL
(\link{monty_dsl}) and also in the related odin DSL.
}
\section{Differences to \code{\link[=D]{D()}}}{
\itemize{
\item We try a little harder to simplify expressions.
\item The distribution functions in the monty DSL (e.g., Poisson)
are (will be) handled specially, allowing substitution of
log-densities and expectations.
\item Once we support array expressions, we will be able to
differentiate through these.
}
}

\section{Roadmap}{
We may need to make this slightly extensible in future, but for
now the set of functions that can be differentiated is closed.
}

\section{Warning}{
The way of accessing distribution support here is peculiar and the
return type unusual.  This is intentional, and we expect a more
conventional interface in the future once this package settles
down.
}

\examples{
d <- monty_differentiation()
d$differentiate(quote(sqrt(sin(x))), "x")
D(quote(sqrt(sin(x))), "x")
}
