% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{monty_dsl_parse_distribution}
\alias{monty_dsl_parse_distribution}
\title{Parse distribution expression}
\usage{
monty_dsl_parse_distribution(expr, name = NULL)
}
\arguments{
\item{expr}{An expression}

\item{name}{Name for the expression, used in constructing messages
that you can use in errors.}
}
\value{
A list; the contents of this are subject to change.
However you can (to a degree) rely on the following elements:
\itemize{
\item \code{name}: The name of the distribution (e.g., \code{Normal}).  This
will be the same as the name of the function called in \code{expr}
\item \code{variant}: The name of the distribution variant, if more than
one is supported.
\item \code{args}: The arguments that you provided, in position-matched
order
\item \code{cpp}: The names of the C++ entrypoint to use.  This is a list
with elements \code{density} and \code{sample} for the log-density and
sampling functions, and \code{NULL} where these do not yet exist.
}

Currently we also include:
\itemize{
\item \code{density}: A function to compute the log-density.  This will
likely change once we support creation of differentiable models
because we will want to do something with the arguments
provided!
\item \code{sample}: A function to sample from the distribution, given (as
a first argument) a rng object (see \link{monty_rng})
}
}
\description{
Parse an expression as if it were a call to one of monty's
distribution functions (e.g., \code{Normal}, \code{Poisson}).  This will
fill in any defaults, disambiguate where multiple
parameterisations of the distribution are available, and provide
links through to the C++ API.  This function is designed for use
from other packages that use monty, and is unlikely to be
useful to most users.
}
