% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-tools.R
\name{monty_flatten_chains}
\alias{monty_flatten_chains}
\title{Flatten chains}
\usage{
monty_flatten_chains(samples)
}
\arguments{
\item{samples}{A \code{monty_samples} object, from running \code{\link[=monty_sample]{monty_sample()}}}
}
\value{
A \code{monty_samples} object with the chains dimension collapsed into
the samples dimension.
}
\description{
Flatten chains in the results of running \code{\link[=monty_sample]{monty_sample()}}, wherein \code{pars},
\code{density} and typically objects in \code{observations} will have the last two
dimensions representing 'samples' and 'chains'. Flattening the chains
results in the chains dimension being collapsed into the samples dimension.
}
