% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{monty_model_direct_sample}
\alias{monty_model_direct_sample}
\title{Directly sample from a model}
\usage{
monty_model_direct_sample(model, rng, named = FALSE)
}
\arguments{
\item{model}{A \link{monty_model} object}

\item{rng}{Random number state, created by \link{monty_rng}.  Use of
an RNG with more than one stream may or may not work as
expected; this is something we need to tidy up (\code{mrc-5292})}

\item{named}{Logical, indicating if the output should be named
using the parameter names.}
}
\value{
A vector or matrix of sampled parameters
}
\description{
Directly sample from a model.  Not all models support this, and an
error will be thrown if it is not possible.
}
\examples{
m <- monty_example("banana")

r <- monty_rng_create()
monty_model_direct_sample(m, r)
monty_model_direct_sample(m, r, named = TRUE)

r <- monty_rng_create(n_streams = 3)
monty_model_direct_sample(m, r)
monty_model_direct_sample(m, r, named = TRUE)
}
