% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{monty_model_split}
\alias{monty_model_split}
\title{Split a combined model}
\usage{
monty_model_split(model, prior_first = FALSE)
}
\arguments{
\item{model}{A combined model}

\item{prior_first}{Logical, indicating that we should require that
the model component that could be the prior is listed first.  If
\code{TRUE} and one component model is not plausibly the prior, we
will error.  See Details for the heuristic used.}
}
\value{
An unnamed list of length 2, being the component models.
If one model might be the prior it will be listed first.
}
\description{
Split a model that has been combined by \code{\link[=monty_model_combine]{monty_model_combine()}} into
its constituent parts.
}
\details{
We assume that a split model can be broken into a "prior" and a
"likelihood" if exactly one model:
\itemize{
\item can be directly sampled from
\item is not stochastic
\item consumes all parameters
}

Typically, it will be the first criterion that will separate a
model into prior and likelihood (if you could sample from your
likelihood, then you would not use a sampler, which is where we
are typically going to perform this action).

If \code{prior_first} is \code{FALSE} we just return the parts.
}
