% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_beta}
\alias{monty_random_beta}
\alias{monty_random_n_beta}
\title{Sample from beta distribution}
\usage{
monty_random_beta(a, b, state)

monty_random_n_beta(n_samples, a, b, state)
}
\arguments{
\item{a, b}{The shape parameters}

\item{state}{The random number state, from \link{monty_rng_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from the beta distribution
}
