% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_beta_binomial_prob}
\alias{monty_random_beta_binomial_prob}
\alias{monty_random_n_beta_binomial_prob}
\alias{monty_random_beta_binomial_ab}
\alias{monty_random_n_beta_binomial_ab}
\title{Sample from beta-binomial distribution}
\usage{
monty_random_beta_binomial_prob(size, prob, rho, state)

monty_random_n_beta_binomial_prob(n_samples, size, prob, rho, state)

monty_random_beta_binomial_ab(size, a, b, state)

monty_random_n_beta_binomial_ab(n_samples, size, a, b, state)
}
\arguments{
\item{size}{The number of trials (zero or more)}

\item{prob}{The mean probability of success on each trial (between
0 and 1)}

\item{rho}{The dispersion parameter (between 0 and 1)}

\item{state}{The random number state, from \link{monty_rng_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}

\item{a}{The first shape parameter (zero or more)}

\item{b}{The second shape parameter (zero or more)}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from a beta-binomial distribution. There are two
parameterisations available: one in terms of probability and
dispersion and the other in terms of two shape parameters.
}
