% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_binomial}
\alias{monty_random_binomial}
\alias{monty_random_n_binomial}
\title{Sample from binomial distribution}
\usage{
monty_random_binomial(size, prob, state)

monty_random_n_binomial(n_samples, size, prob, state)
}
\arguments{
\item{size}{The number of trials}

\item{prob}{The probability of success on each trial}

\item{state}{The random number state, from \link{monty_rng_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from the binomial distribution
}
\examples{
state <- monty_random_create()
monty_random_binomial(10, 0.3, state)
table(monty_random_n_binomial(2000, 10, 0.3, state))
}
