% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_create}
\alias{monty_random_create}
\title{Create a monty random number genreator}
\usage{
monty_random_create(
  n_streams = 1L,
  seed = NULL,
  n_threads = 1L,
  deterministic = FALSE,
  preserve_stream_dimension = FALSE
)
}
\arguments{
\item{n_streams}{The number of streams to create (see Details)}

\item{seed}{The initial seed of the random number generator.  This
can be \code{NULL}, in which case we seed the generator from R's
random number state (meaning that we respond to \code{set.seed} as
one would expect).  Alternatively, you can provide an integer
here, but this should be used sparingly and primarily for
testing.}

\item{n_threads}{The number of threads to use, if OpenMP is
enabled.}

\item{deterministic}{Logical, indicating if we should use
"deterministic" mode where distributions return their
expectations and the state is never changed.}

\item{preserve_stream_dimension}{Logical, indicating if the stream
dimension should be preserved in the case where \code{n_streams} is 1
and the multiple-sample functions are used.  Set this to \code{TRUE}
to ensure that the rank of the result does not change with the
number of streams (see Details).}
}
\value{
An object of class \code{monty_random_state}, which can be
passed as the \code{state} argument to random-number producing
functions, such as \link{monty_random_real}
}
\description{
Create a monty random number generator.  This allows you to sample
random numbers from the same random number algorithms as monty
provides via C++ to \href{https://mrc-ide.github.io/dust2}{dust2}, and
which it uses within its samplers and filters.  This function
creates the internal state, and will be passed to actual
generation functions \verb{monty_random_*}, such as
\code{\link[=monty_random_real]{monty_random_real()}}.
}
\details{
Monty's random number generation is very different to that in base
R.  We have the concept of "streams" of random numbers, with a
generator having 1 or many streams.  Each stream is statistically
independent, and can be sampled from simultaneously.  When you use
any of the random number functions from R, you will draw one
number \emph{per stream}.

Because the random number state can have multiple streams, and we
return a vector, we have a separate set of functions where
multiple numbers are requested \emph{per stream}; these are all
prefixed \code{monty_random_n_} (e.g., \code{\link[=monty_random_n_real]{monty_random_n_real()}}).  These
will return a matrix where you have used multiple streams, with
each column representing a stream.  If you have a single stream
and you set \code{preserve_stream_dimension} to \code{FALSE} then we will
drop this dimension and return a matrix.
}
\examples{
state <- monty_random_create()
state

monty_random_real(state)
}
