% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_exponential_rate}
\alias{monty_random_exponential_rate}
\alias{monty_random_n_exponential_rate}
\alias{monty_random_exponential_mean}
\alias{monty_random_n_exponential_mean}
\title{Sample from exponential distribution}
\usage{
monty_random_exponential_rate(rate, state)

monty_random_n_exponential_rate(n_samples, rate, state)

monty_random_exponential_mean(mean, state)

monty_random_n_exponential_mean(n_samples, size, mean)
}
\arguments{
\item{rate}{The rate of the exponential}

\item{state}{The random number state, from \link{monty_rng_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}

\item{mean}{The mean of the exponential distribution (i.e., \code{1 / rate})}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from an exponential distribution.  There are two
parameterisations here, one in terms of the rate of the
exponential, and one in terms of the mean (or scale).
}
\examples{
state <- monty_rng_create()
monty_random_exponential_rate(0.2, state)
summary(monty_random_n_exponential_rate(2000, 0.2, state))
}
