% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_gamma_scale}
\alias{monty_random_gamma_scale}
\alias{monty_random_n_gamma_scale}
\alias{monty_random_gamma_rate}
\alias{monty_random_n_gamma_rate}
\title{Sample from a gamma distribution.  There are two parameterisations
here, one in terms of rate, and one in terms of scale.}
\usage{
monty_random_gamma_scale(shape, scale, state)

monty_random_n_gamma_scale(n_samples, shape, scale, state)

monty_random_gamma_rate(shape, rate, state)

monty_random_n_gamma_rate(n_samples, shape, rate, state)
}
\arguments{
\item{shape}{Shape}

\item{scale}{Scale
'}

\item{state}{The random number state, from \link{monty_random_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}

\item{n_threads}{Number of threads to use; see Details}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from gamma distribution
}
