% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_negative_binomial_prob}
\alias{monty_random_negative_binomial_prob}
\alias{monty_random_n_negative_binomial_prob}
\alias{monty_random_negative_binomial_mu}
\alias{monty_random_n_negative_binomial_mu}
\title{Sample from negative binomial distribution}
\usage{
monty_random_negative_binomial_prob(size, prob, state)

monty_random_n_negative_binomial_prob(n_samples, size, prob, state)

monty_random_negative_binomial_mu(size, mu, state)

monty_random_n_negative_binomial_mu(n_samples, size, mu, state)
}
\arguments{
\item{size}{The target number of successful trials
(zero or more)}

\item{prob}{The probability of success on each trial (between 0
and 1)}

\item{state}{The random number state, from \link{monty_rng_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}

\item{mu}{The mean (zero or more)}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from a negative binomial distribution. There are two
parameterisations available: one in based on the success probability and the
other based on the mean.
}
