% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_real}
\alias{monty_random_real}
\alias{monty_random_n_real}
\title{Sample from Uniform(0, 1)}
\usage{
monty_random_real(state)

monty_random_n_real(n_samples, state)
}
\arguments{
\item{state}{The random number state, from \link{monty_random_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Generate a random number uniformly sampled on the range 0 to 1;
this is the most basic of all random number functions in monty and
all other algorithms are composed from this.  Quite often, you
will want a number on \link{0, 1} (e.g., for a Bernoulli trial), and
this function is the most efficient way of generating one.
}
\examples{
state <- monty_random_create()
monty_random_real(state)
monty_random_n_real(5, state)
}
