% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_state}
\alias{monty_random_state}
\alias{monty_random_set_state}
\title{Get and set random number state}
\usage{
monty_random_state(state)

monty_random_set_state(value, state)
}
\arguments{
\item{state}{The random number state, from \link{monty_random_create}}

\item{value}{A vector of raw values, typically the result of
exporting a random state with \code{monty_random_state()}}
}
\value{
A vector of raws
}
\description{
Get and set internal random number state
}
\examples{
s1 <- monty_random_create()
r1 <- monty_random_state(s1)

s2 <- monty_random_create(seed = r1)
identical(r1, monty_random_state(s2))
monty_random_real(s1)
monty_random_real(s2)

monty_random_set_state(s1, r1)
monty_random_real(s1)
monty_random_real(s1)
monty_random_real(s2)
}
