% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_truncated_normal}
\alias{monty_random_truncated_normal}
\alias{monty_random_n_truncated_normal}
\title{Sample from truncated normal}
\usage{
monty_random_truncated_normal(mean, sd, min, max, state)

monty_random_n_truncated_normal(n_samples, mean, sd, min, max, state)
}
\arguments{
\item{mean}{The mean of the \strong{parent} (untruncated) normal
distribution (this is not necessarily the mean of the truncated
distribution, unless the \code{min} and \code{max} are symmetrically
placed around \code{mean})}

\item{sd}{The standard deviation of the \strong{parent} distribution
(this is not the same as the standard deviation of the truncated
distribution with finite bounds).}

\item{min}{The lower bound (can be \code{-Inf}).}

\item{max}{The upper bound (can be \code{Inf}).}

\item{state}{The random number state, from \link{monty_random_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from a truncated normal distribution
}
