% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_uniform}
\alias{monty_random_uniform}
\alias{monty_random_n_uniform}
\title{Sample from uniform distribution}
\usage{
monty_random_uniform(min, max, state)

monty_random_n_uniform(n_samples, min, max, state)
}
\arguments{
\item{min}{The minimum value of the uniform distribution}

\item{max}{The maximum value of the uniform distribution}

\item{state}{The random number state, from \link{monty_rng_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from a uniform distribution
}
