% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_zi_poisson}
\alias{monty_random_zi_poisson}
\alias{monty_random_n_zi_poisson}
\title{Sample from zero-inflated Poisson distribution}
\usage{
monty_random_zi_poisson(pi0, lambda, state)

monty_random_n_zi_poisson(n_samples, pi0, lambda, state)
}
\arguments{
\item{pi0}{The probability of extra zeros}

\item{lambda}{The mean (zero or more, length 1 or n) of the Poisson part of
the distribution. Only valid for lambda <= 10^7. Note that the mean of the
zero-inflated Poisson distribution will actually be \code{(1 - pi0) * lambda}}

\item{state}{The random number state, from \link{monty_rng_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from the zero-inflated Poisson distribution. With probability \code{pi0} a
0 is drawn, otherwise it draws from the corresponding standard Poisson
distribution (which may also result in a 0 being drawn).
}
