% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_distributed.R
\name{monty_rng_distributed_state}
\alias{monty_rng_distributed_state}
\alias{monty_rng_distributed_pointer}
\title{Create a set of distributed seeds}
\usage{
monty_rng_distributed_state(
  seed = NULL,
  n_streams = 1L,
  n_nodes = 1L,
  algorithm = "xoshiro256plus"
)

monty_rng_distributed_pointer(
  seed = NULL,
  n_streams = 1L,
  n_nodes = 1L,
  algorithm = "xoshiro256plus"
)
}
\arguments{
\item{seed}{Initial seed to use. As for \link{monty_rng}, this can
be \code{NULL} (create a seed using R's generators), an integer or a
raw vector of appropriate length.}

\item{n_streams}{The number of streams to create per node.}

\item{n_nodes}{The number of separate seeds to create. Each will
be separated by a "long jump" for your generator.}

\item{algorithm}{The name of an algorithm to use.}
}
\value{
A list of either raw vectors (for
\code{monty_rng_distributed_state}) or of \link{monty_rng_pointer}
objects (for \code{monty_rng_distributed_pointer})
}
\description{
Create a set of initial random number seeds suitable for using
within a distributed context (over multiple processes or nodes) at
a level higher than a single group of synchronised threads.
}
\details{
See \code{vignette("rng_distributed")} for a proper introduction to
these functions.
}
\examples{
monty::monty_rng_distributed_state(n_nodes = 2)
monty::monty_rng_distributed_pointer(n_nodes = 2)
}
