% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_rng_jump}
\alias{monty_rng_jump}
\alias{monty_rng_long_jump}
\title{Jump random number state}
\usage{
monty_rng_jump(state, n = 1)

monty_rng_long_jump(state, n = 1)
}
\arguments{
\item{state}{Either a \code{monty_rng_state} object (created via
\link{monty_rng_create}) or a raw vector suitable for creating
one.}

\item{n}{The number of jumps to take (integer, 1 or more)}
}
\value{
The \code{monty_rng_state} object (invisibly, modified in
place) or a raw vector, matching the input argument \code{state}
(visibly).
}
\description{
Jump random number state.  There are two "lengths" of jumps; a
normal jump and a long jump.  The normal jump is the distance
between streams within a random number state, so if you have a
multi-stream rng this shifts states left.  The long jump is used
to create distributed states.  We will properly explain all this
once the interface stabilises.
}
