% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_pointer.R
\name{monty_rng_pointer}
\alias{monty_rng_pointer}
\title{Create pointer to random number generator stream}
\description{
This function exists to support use from other
packages that wish to use monty's random number support, and
creates an opaque pointer to a set of random number streams.
}
\examples{
monty::monty_rng_pointer$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{algorithm}}{The name of the generator algorithm used (read-only)}

\item{\code{n_streams}}{The number of streams of random numbers provided
(read-only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-monty_rng_pointer-new}{\code{monty_rng_pointer$new()}}
\item \href{#method-monty_rng_pointer-sync}{\code{monty_rng_pointer$sync()}}
\item \href{#method-monty_rng_pointer-state}{\code{monty_rng_pointer$state()}}
\item \href{#method-monty_rng_pointer-is_current}{\code{monty_rng_pointer$is_current()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monty_rng_pointer-new"></a>}}
\if{latex}{\out{\hypertarget{method-monty_rng_pointer-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{monty_rng_pointer} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{monty_rng_pointer$new(
  seed = NULL,
  n_streams = 1L,
  long_jump = 0L,
  algorithm = "xoshiro256plus"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{The random number seed to use (see \link{monty_rng}
for details)}

\item{\code{n_streams}}{The number of independent random number streams to
create}

\item{\code{long_jump}}{Optionally an integer indicating how many
"long jumps" should be carried out immediately on creation.
This can be used to create a distributed parallel random number
generator (see \link{monty_rng_distributed_state})}

\item{\code{algorithm}}{The random number algorithm to use. The default is
\code{xoshiro256plus} which is a good general choice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monty_rng_pointer-sync"></a>}}
\if{latex}{\out{\hypertarget{method-monty_rng_pointer-sync}{}}}
\subsection{Method \code{sync()}}{
Synchronise the R copy of the random number state.
Typically this is only needed before serialisation if you have
ever used the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{monty_rng_pointer$sync()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monty_rng_pointer-state"></a>}}
\if{latex}{\out{\hypertarget{method-monty_rng_pointer-state}{}}}
\subsection{Method \code{state()}}{
Return a raw vector of state. This can be used to
create other generators with the same state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{monty_rng_pointer$state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-monty_rng_pointer-is_current"></a>}}
\if{latex}{\out{\hypertarget{method-monty_rng_pointer-is_current}{}}}
\subsection{Method \code{is_current()}}{
Return a logical, indicating if the random number
state that would be returned by \code{state()} is "current" (i.e., the
same as the copy held in the pointer) or not. This is \code{TRUE} on
creation or immediately after calling \verb{$sync()} or \verb{$state()}
and \code{FALSE} after any use of the pointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{monty_rng_pointer$is_current()}\if{html}{\out{</div>}}
}

}
}
