% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_rng_state}
\alias{monty_rng_state}
\alias{monty_rng_set_state}
\title{Get and set random number state}
\usage{
monty_rng_state(state)

monty_rng_set_state(value, state)
}
\arguments{
\item{state}{The random number state, from \link{monty_rng_create}}

\item{value}{A vector of raw values, typically the result of
exporting a random state with \code{monty_rng_state()}}
}
\value{
A vector of raws
}
\description{
Get and set internal random number state
}
\examples{
s1 <- monty_rng_create()
r1 <- monty_rng_state(s1)

s2 <- monty_rng_create(seed = r1)
identical(r1, monty_rng_state(s2))
monty_random_real(s1)
monty_random_real(s2)

monty_rng_set_state(r1, s1)
monty_random_real(s1)
monty_random_real(s1)
monty_random_real(s2)
}
