% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{monty_runner_serial}
\alias{monty_runner_serial}
\title{Run MCMC chain in series}
\usage{
monty_runner_serial(progress = NULL)
}
\arguments{
\item{progress}{Optional logical, indicating if we should print a
progress bar while running.  If \code{NULL}, we use the value of the
option \code{monty.progress} if set, otherwise we show the progress
bar (as it is typically wanted).  Alternatively, you can provide
a string indicating the progress bar type.  Options are \code{fancy}
(equivalent to \code{TRUE}), \code{none} (equivalent to \code{FALSE}) and
\code{simple} (a very simple text-mode progress indicator designed
play nicely with logging; it does not use special codes to clear
the line).}
}
\value{
A runner of class \code{monty_runner} that can be passed to
\code{\link[=monty_sample]{monty_sample()}}
}
\description{
Run MCMC chains in series (one after another).  This is the
simplest chain runner, and the default used by \code{\link[=monty_sample]{monty_sample()}}.
It has nothing that can be configured (yet).
}
