% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner-simultaneous.R
\name{monty_runner_simultaneous}
\alias{monty_runner_simultaneous}
\title{Run MCMC chains simultaneously}
\usage{
monty_runner_simultaneous(progress = NULL)
}
\arguments{
\item{progress}{Optional logical, indicating if we should print a
progress bar while running.  If \code{NULL}, we use the value of the
option \code{monty.progress} if set, otherwise we show the progress
bar (as it is typically wanted).  Alternatively, you can provide
a string indicating the progress bar type.  Options are \code{fancy}
(equivalent to \code{TRUE}), \code{none} (equivalent to \code{FALSE}) and
\code{simple} (a very simple text-mode progress indicator designed
play nicely with logging; it does not use special codes to clear
the line).}
}
\value{
A runner of class \code{monty_runner} that can be passed to
\code{\link[=monty_sample]{monty_sample()}}
}
\description{
Run chains \emph{simultaneously}.  This differs from
\link{monty_runner_parallel}, which runs chains individually in
parallel by working with models that can evaluate multiple
densities at the same time.  There are situations where this might
be faster than running in parallel, but primarily this exists so
that we can see that samplers can work with multiple samples at
once.
}
\examples{
m <- monty_example("banana")
s <- monty_sampler_random_walk(vcv = diag(2) * 0.01)
r <- monty_runner_simultaneous()
samples <- monty_sample(m, s, 200, runner = r)
}
