% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{monty_sample_continue}
\alias{monty_sample_continue}
\title{Continue sampling}
\usage{
monty_sample_continue(samples, n_steps, restartable = FALSE, runner = NULL)
}
\arguments{
\item{samples}{A \code{monty_samples} object created by
\code{\link[=monty_sample]{monty_sample()}}}

\item{n_steps}{The number of new steps to run}

\item{restartable}{Logical, indicating if the chains should be
restartable.  This will add additional data to the chains
object.}

\item{runner}{Optionally, a runner for your chains.  The default
is to continue with the backend that you used to start the
chains via \link{monty_sample} (or on the previous restart with
this function).  You can use this argument to change the runner,
which might be useful if transferring a pilot run from a
high-resource environment to a lower-resource environment.  If
given, must be an \code{monty_runner} object such as
\link{monty_runner_serial} or \link{monty_runner_parallel}.  You can
use this argument to change the configuration of a runner, as
well as the type of runner (e.g., changing the number of
allocated cores).}
}
\value{
A list of parameters and densities
}
\description{
Continue (restart) chains started by \link{monty_sample}.  Requires
that the original chains were run with \code{restartable = TRUE}.
Running chains this way will result in the final state being
exactly the same as running for the total (original + continued)
number of steps in a single push.
}
