% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-manual.R
\name{monty_sample_manual_cleanup}
\alias{monty_sample_manual_cleanup}
\title{Clean up samples}
\usage{
monty_sample_manual_cleanup(path)
}
\arguments{
\item{path}{The path used in the call to
\link{monty_sample_manual_prepare}}
}
\value{
Nothing, called for side effects only.
}
\description{
Clean up after manual sampling.  This is essentially a safe
version of deleting the directory (e.g, \code{unlink(path, recursive = TRUE)}) which checks that the directory really was used for
sampling and that it does not contain anything else unexpected.
}
\examples{

model <- monty_example("banana")
sampler <- monty_sampler_random_walk(vcv = diag(2) * 0.05)
path <- tempfile()
monty_sample_manual_prepare(model, sampler, 100, path)
monty_sample_manual_info(path)

# Run the (single) chain
monty_sample_manual_run(1, path)
monty_sample_manual_info(path)

# Collect the results
monty_sample_manual_collect(path)

# Clean up samples
monty_sample_manual_cleanup(path)
}
