% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-manual.R
\name{monty_sample_manual_collect}
\alias{monty_sample_manual_collect}
\title{Collect manually run samples}
\usage{
monty_sample_manual_collect(
  path,
  samples = NULL,
  restartable = FALSE,
  append = TRUE
)
}
\arguments{
\item{path}{The path used in the call to
\link{monty_sample_manual_prepare}}

\item{samples}{Samples from the parent run.  You need to provide
these where \code{save_samples} was set to anything other than "value"}

\item{restartable}{Logical, indicating if the chains should be
restartable.  This will add additional data to the chains
object.  Note that this is controlled at chain collection and
not creation.}
}
\value{
A \code{monty_samples} object.
}
\description{
Collect samples from chains that have been run with
\link{monty_sample_manual_prepare} and \link{monty_sample_manual_run}.  If
any chain has not completed, we will error.
}
\examples{

model <- monty_example("banana")
sampler <- monty_sampler_random_walk(vcv = diag(2) * 0.05)
path <- tempfile()
monty_sample_manual_prepare(model, sampler, 100, path)
monty_sample_manual_info(path)

# Run the (single) chain
monty_sample_manual_run(1, path)
monty_sample_manual_info(path)

# Collect the results
monty_sample_manual_collect(path)

# Clean up samples
monty_sample_manual_cleanup(path)
}
