% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-manual.R
\name{monty_sample_manual_info}
\alias{monty_sample_manual_info}
\title{Get information about manually scheduled samples}
\usage{
monty_sample_manual_info(path)
}
\arguments{
\item{path}{The path used in the call to
\link{monty_sample_manual_prepare}}
}
\value{
Invisibly, a logical vector, \code{TRUE} for completed chains
and \code{FALSE} for incomplete chains.
}
\description{
Get information about the status of manually scheduled samples.
}
\examples{

model <- monty_example("banana")
sampler <- monty_sampler_random_walk(vcv = diag(2) * 0.05)
path <- tempfile()
monty_sample_manual_prepare(model, sampler, 100, path)
monty_sample_manual_info(path)

# Run the (single) chain
monty_sample_manual_run(1, path)
monty_sample_manual_info(path)

# Collect the results
monty_sample_manual_collect(path)

# Clean up samples
monty_sample_manual_cleanup(path)
}
