% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-manual.R
\name{monty_sample_manual_prepare_continue}
\alias{monty_sample_manual_prepare_continue}
\title{Prepare to continue sampling from a model, with manual chain
orchestration.  This function is to \link{monty_sample_continue} what
\link{monty_sample_manual_prepare} is to \link{monty_sample}.  The original
set of samples do not need to have been run manually.}
\usage{
monty_sample_manual_prepare_continue(
  samples,
  n_steps,
  path,
  save_samples = "hash"
)
}
\arguments{
\item{n_steps}{The number of steps to run the sampler for.}

\item{path}{The path to write inputs and outputs to.  This should
be a path to a directory which does not yet exist, or which is
empty; we will create one here.  The contents of this directory
are managed by \code{monty} and the names and contents of files here
are an implementation detail and should not be relied on.
Calling \code{monty_sample_manual_cleanup()} will delete the
directory in its entirety.  Be aware that if you use
\code{\link[=tempfile]{tempfile()}} here (which can be a reasonable choice!) that this
path will be deleted when your R process ends, so if using these
the process calling \code{monty_sample_manual_prepare} should outlive
running all sampling.}

\item{save_samples}{Control over saving samples into the inputs.
The choices here are \code{hash} (the default) where we save a hash
and validate that in \link{monty_sample_manual_collect}, \code{value}
where the samples are themselves saved and you can omit the
\code{samples } argument to \link{monty_sample_manual_collect}, or
\code{nothing}, in which we save nothing, and you just have to get it
right.}
}
\value{
Invisibly, the path used to store files (the same as the
value of the \code{path} argument)
}
\description{
Prepare to continue sampling from a model, with manual chain
orchestration.  This function is to \link{monty_sample_continue} what
\link{monty_sample_manual_prepare} is to \link{monty_sample}.  The original
set of samples do not need to have been run manually.
}
