% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-manual.R
\name{monty_sample_manual_run}
\alias{monty_sample_manual_run}
\title{Run sample with manual scheduling}
\usage{
monty_sample_manual_run(chain_id, path, progress = NULL)
}
\arguments{
\item{chain_id}{The id for the chain to run, an integer.  If you
provide an integer that does not correspond to a chain in 1 to
\code{n_chains} (where \code{n_chains} was the argument passed to
\code{monty_sample_manual_prepare} it is an error.}

\item{path}{The path used in the call to
\link{monty_sample_manual_prepare}}

\item{progress}{Optional logical, indicating if we should print a
progress bar while running.  If \code{NULL}, we use the value of the
option \code{monty.progress} if set, otherwise we show the
progress bar (as it is typically wanted).  The progress bar
itself responds to cli's options; in particular
\code{cli.progress_show_after} and \code{cli.progress_clear} will affect
your experience.}
}
\description{
Run a chain that was prepared using \link{monty_sample_manual_prepare}.
}
\section{Warning:}{
There is no lock mechanism; you can start a single chain many
times.  Don't do this.
}

\examples{

model <- monty_example("banana")
sampler <- monty_sampler_random_walk(vcv = diag(2) * 0.05)
path <- tempfile()
monty_sample_manual_prepare(model, sampler, 100, path)
monty_sample_manual_info(path)

# Run the (single) chain
monty_sample_manual_run(1, path)
monty_sample_manual_info(path)

# Collect the results
monty_sample_manual_collect(path)

# Clean up samples
monty_sample_manual_cleanup(path)
}
