% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{monty_sampler_properties}
\alias{monty_sampler_properties}
\title{Describe sampler properties}
\usage{
monty_sampler_properties(
  has_state = NULL,
  restartable = NULL,
  allow_multiple_parameters = FALSE,
  requires_gradient = FALSE,
  requires_allow_multiple_parameters = FALSE,
  requires_deterministic = FALSE
)
}
\arguments{
\item{has_state}{Optional logical, indicating if the sampler has
state.  This is optional because presence of the state function
\code{state_dump} implies this.}

\item{restartable}{Optional logical, indicating if your sampler
can be restarted.  If \code{FALSE}, then users cannot use
\code{restartable = TRUE} from \code{monty_sample()} (and therefore cannot
use \code{monty_continue}).  This is optional because the presence of
the state function \code{state_restore} implies this.}

\item{allow_multiple_parameters}{Logical, indicating if your
sampler can accept a matrix of parameters in order to run
multiple chains at once (e.g., with the
\link{monty_runner_simultaneous} runner, or as part of a parallel
tempering scheme with \link{monty_sampler_parallel_tempering}).}

\item{requires_gradient}{Logical, indicating if the model must
provide a gradient in order to be used with this sampler.}

\item{requires_allow_multiple_parameters}{Logical, indicating if
the model must be able to accept multiple parameters.  This is
different to \code{allow_multiple_parameters}, which concerns if the
\emph{sampler} is able to process multiple parameter sets at once.
For example, \link{monty_sampler_parallel_tempering} sets
\code{allow_multiple_parameters} to \code{FALSE} but
\code{requires_allow_multiple_parameters} to \code{FALSE}, while
\link{monty_sampler_random_walk} sets the opposite!}

\item{requires_deterministic}{Logical, indicating if the model
must be deterministic in order to be used with this sampler.}
}
\value{
A \code{monty_sampler_properties} object, which should not be
modified.
}
\description{
Describe properties of a sampler.  This is used from
\link{monty_sampler} to advertise what your sampler does about state,
what it requires from the runner and from the model, so that monty
can ensure that it is only used where it is appropriate.y
}
\examples{
monty_sampler_properties()
}
