% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-tools.R
\name{monty_unflatten_chains}
\alias{monty_unflatten_chains}
\title{Unflatten chains}
\usage{
monty_unflatten_chains(samples)
}
\arguments{
\item{samples}{A \code{monty_samples} object, that has been run through
\code{\link[=monty_flatten_chains]{monty_flatten_chains()}}}
}
\value{
A \code{monty_samples} object with the chains dimension restored.
}
\description{
Unflatten chains in \code{monty_samples} object that has previously had the
chains flattened with \code{\link[=monty_flatten_chains]{monty_flatten_chains()}}, reversing the effects of
that function.
}
