% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_goat_door.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A vector of "goat" and "car" representing the game}

\item{a.pick}{An integer representing the number of selected door}
}
\value{
An integer representing the number of the opened door
}
\description{
This function takes the 'game' and the 'a.pick' (selected door) as inputs.
If the selected door is a car, it randomly selects one of the other two doors, which are goats
If the selected door is a goat, it opens the other door which is also a goat.
It returns the number of the opened door
}
\examples{
open_goat_door(c("goat", "car", "goat"), 2)

}
