% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{with_trace_random}
\alias{with_trace_random}
\title{Trace random number calls}
\usage{
with_trace_random(code)
}
\arguments{
\item{code}{Code to run with tracing on}
}
\value{
The result of evaluating \code{code}
}
\description{
Trace calls to R's random-number-generating functions, to detect
unexpected use of random number generation outside of mcstate2's
control.
}
