% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mcstate_model_direct_sample}
\alias{mcstate_model_direct_sample}
\title{Directly sample from a model}
\usage{
mcstate_model_direct_sample(model, rng, named = FALSE)
}
\arguments{
\item{model}{An \link{mcstate_model} object}

\item{rng}{Random number state, created by \link{mcstate_rng}.  Use of
an RNG with more than one stream may or may not work as
expected; this is something we need to tidy up (mrc-5292)}

\item{named}{Logical, indicating if the output should be named
using the parameter names.}
}
\value{
A vector or matrix of sampled parameters
}
\description{
Directly sample from a model.  Not all models suport this, and an
error will be thrown if it is not possible.
}
