% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mcstate_model_gradient}
\alias{mcstate_model_gradient}
\title{Compute gradient of log density}
\usage{
mcstate_model_gradient(model, parameters, named = FALSE)
}
\arguments{
\item{model}{An \link{mcstate_model} object}

\item{parameters}{A vector or matrix of parameters}

\item{named}{Logical, indicating if the output should be named
using the parameter names.}
}
\value{
A vector or matrix of gradients
}
\description{
Compute the gradient of log density (which is returned by
\link{mcstate_model_density}) with respect to parameters.  Not all models
support this, and an error will be thrown if it is not possible.
}
\seealso{
\link{mcstate_model_density} for log density, and
\link{mcstate_model_direct_sample} to sample from a model
}
