% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{monty_example}
\alias{monty_example}
\title{Example models}
\usage{
monty_example(name, ...)
}
\arguments{
\item{name}{Name of the example, as a string.  See Details for
supported models.}

\item{...}{Optional parameters that are passed to create the
model.  All models can be created with no additional parameters,
but you can tweak their behaviour by passing named parameters
here.  See Details.}
}
\value{
A \link{monty_model} object
}
\description{
Load example models from monty.  These models exist so that we can
create (hopefully) interesting examples in the documentation
without them becoming overwhelming.  You should probably not use
these for anything other than exploring the package.
}
\section{Supported models}{
\subsection{\code{bananna}}{

The banana model is a two-dimensional banana-shaped function,
picked because it is quite annoying to sample from directly.  The
model has two parameters \code{alpha} and \code{beta} and is based on two
successive draws, one condiional on the other:

[
\beta \sim Normal(1, 0) \\
\alpha \sim Normal(\beta^2, \sigma)
]

You can vary \code{sigma} for this model on creation, the default is 0.5
}
}

\examples{
monty_example("banana")
}
