% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-function.R
\name{monty_model_function}
\alias{monty_model_function}
\title{Create \code{monty_model} from a function computing density}
\usage{
monty_model_function(density, packer = NULL, fixed = NULL)
}
\arguments{
\item{density}{A function to compute log density.  It can take any
number of parameters}

\item{packer}{Optionally, a \link{monty_packer} object to control how
your function parameters are packed into a numeric vector.  You
can typically omit this if all the arguments to your functions
are present in your numeric vector and if they are all scalars.}

\item{fixed}{Optionally, a named list of fixed values to
substitute into the call to \code{density}.  This cannot be used in
conjunction with \code{packer} (you should use the \code{fixed} argument
to \code{monty_packer} instead).}
}
\value{
A \link{monty_model} object that computes log density with the
provided \code{density} function, given a numeric vector argument
representing all parameters.
}
\description{
Create a \link{monty_model} from a function that computes density.
This allows use of any R function as a simple monty model.  If you
need advanced model features, then this interface may not suit you
and you may prefer to use \link{monty_model} directly.
}
\details{
This interface will expand in future versions of monty to support
gradients, stochastic models, parameter groups and simultaneous
calculation of density.
}
