% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{monty_model_gradient}
\alias{monty_model_gradient}
\title{Compute gradient of log density}
\usage{
monty_model_gradient(model, parameters, augmented_data = NULL, named = FALSE)
}
\arguments{
\item{model}{A \link{monty_model} object}

\item{parameters}{A vector or matrix of parameters}

\item{named}{Logical, indicating if the output should be named
using the parameter names.}
}
\value{
A vector or matrix of gradients
}
\description{
Compute the gradient of log density (which is returned by
\link{monty_model_density}) with respect to parameters.  Not all models
support this, and an error will be thrown if it is not possible.
}
\examples{
m <- monty_example("banana")
# Global maximum at (0, 0), and gradient is zero there:
monty_model_density(m, c(0, 0))
monty_model_gradient(m, c(0, 0))

# Nonzero gradient away from the origin:
monty_model_gradient(m, c(0.4, 0.2))
}
\seealso{
\link{monty_model_density} for log density, and
\link{monty_model_direct_sample} to sample from a model
}
