% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{monty_model_split}
\alias{monty_model_split}
\title{Split a combined model}
\usage{
monty_model_split(model, prior_first = FALSE)
}
\arguments{
\item{model}{A combined model}

\item{prior_first}{Logical, indicating that we should require that
the model component that could be the prior is listed first.  If
\code{TRUE} and one component model is not plausibly the prior, we
will error.  See Details for the heuristic used.}
}
\value{
An unnamed list of length 2, being the component models.
If one model might be the prior it will be listed first.
}
\description{
Split a model that has been combined by \code{\link[=monty_model_combine]{monty_model_combine()}} into
its constituent parts.
}
