% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_cauchy}
\alias{monty_random_cauchy}
\title{Sample from Cauchy distribution}
\usage{
monty_random_cauchy(location, scale, state)
}
\arguments{
\item{location}{Location of the distribution (the same as the
median and mode)}

\item{scale}{A scale parameter which specifies the half-width at
half-maximum (HWHM)}

\item{state}{The random number state, from \link{monty_random_create}}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from the Cauchy distribution
}
