% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{monty_random_zi_negative_binomial_prob}
\alias{monty_random_zi_negative_binomial_prob}
\alias{monty_random_n_zi_negative_binomial_prob}
\alias{monty_random_zi_negative_binomial_mu}
\alias{monty_random_n_zi_negative_binomial_mu}
\title{Sample from zero-inflated negative binomial distribution}
\usage{
monty_random_zi_negative_binomial_prob(pi0, size, prob, state)

monty_random_n_zi_negative_binomial_prob(n_samples, pi0, size, prob, state)

monty_random_zi_negative_binomial_mu(pi0, size, mu, state)

monty_random_n_zi_negative_binomial_mu(n_samples, pi0, size, mu, state)
}
\arguments{
\item{pi0}{The probability of excess zeros}

\item{size}{The target number of successful trials
(zero or more)}

\item{prob}{The probability of success on each trial (between 0
and 1)}

\item{state}{The random number state, from \link{monty_rng_create}}

\item{n_samples}{The number of samples to take, \strong{per stream}.
When using the multiple-sample interface, all other parameters
are held constant (per stream).}

\item{mu}{The mean (zero or more) of the negative binomial part of the
distribution}
}
\value{
A vector of random numbers, the same length as the number
of streams in \code{state}.
}
\description{
Sample from a zero-inflated negative binomial distribution
}
