% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner-simultaneous.R
\name{monty_runner_simultaneous}
\alias{monty_runner_simultaneous}
\title{Run MCMC chains simultaneously}
\usage{
monty_runner_simultaneous(progress = NULL)
}
\arguments{
\item{progress}{Optional logical, indicating if we should print a
progress bar while running.  If \code{NULL}, we use the value of the
option \code{monty.progress} if set, otherwise we show the
progress bar (as it is typically wanted).  The progress bar
itself responds to cli's options; in particular
\code{cli.progress_show_after} and \code{cli.progress_clear} will affect
your experience.}
}
\value{
A runner of class \code{monty_runner} that can be passed to
\code{\link[=monty_sample]{monty_sample()}}
}
\description{
Run chains \emph{simultaneously}.  This differs from
\link{monty_runner_parallel}, which runs chains individually in
parallel by working with models that can evaluate multiple
densities at the same time.  There are situations where this might
be faster than running in parallel, but primarily this exists so
that we can see that samplers can work with multiple samples at
once.
}
\examples{
m <- monty_example("banana")
s <- monty_sampler_random_walk(vcv = diag(2) * 0.01)
r <- monty_runner_simultaneous()
samples <- monty_sample(m, s, 200, runner = r)
}
